Update drawre function
This commit is based on a request of Sidney Bush; a Ph.D. student from the University of Colorado Boulder. She used the drawre function by using a time series and value for the  Thres parameter where the number of events below that threshold is zero (a relatively uncommon scenario). Since the prior code of the drawre function did not consider that scenario, it shows an error. Now, the drawre function has been updated to include that scenario. A conditional statement to identify date gaps has also been included.